requirements.txt
sentence-transformers==2.7.0
scikit-learn==1.5.1
numpy>=1.24
PyYAML>=6.0.1
joblib>=1.4.2
ollama>=0.3.0
rank-bm25>=0.2.2
Flask>=3.0.0
flask-cors>=4.0.0
